refactor: nest operator grouping under new `terms` module
Renames the `qiskit_fermions.operators.grouping` module to
`qiskit_fermions.operators.terms.grouping` across the core, cext, and
pyext crates as well as the Python package, docs, and tests. This is a
pure rename with no behavioral changes.

The motivation is that grouping is really one instance of a broader
family of operations that partition an operator's terms based on their
structure (e.g. filtering out bitstring-invariant terms for SqDRIFT,
splitting pairing from number-conserving sectors, k-body truncation).
Rather than let the narrowly-named `grouping` module accrete unrelated
routines, the new `terms` umbrella names the concept those routines
share and gives future term-partitioning functionality a coherent home.
Doing this now is cheap because none of these modules have been released
or are considered stable yet.

This commit was assisted by Claude Opus 4.8
