+++ crates/cext/src/operators/terms/mod.rs
@@ -10,7 +10,4 @@
 // copyright notice, and modified files need to carry a notice indicating
 // that they have been altered from the originals.
 
-pub mod fermion_operator;
-pub mod library;
-pub mod majorana_operator;
-pub mod terms;
+pub mod grouping;
