+++ crates/pyext/src/operators/edge_vertex_operator.rs
@@ -839,7 +839,7 @@ impl PyEdgeVertexOperator {
 
     /// An optional vector of `group indices` for each term.
     ///
-    /// For more information refer to the :mod:`~qiskit_fermions.operators.grouping` module.
+    /// For more information refer to the :mod:`~qiskit_fermions.operators.terms.grouping` module.
     #[getter]
     pub fn get_groups(&self) -> Option<Vec<u32>> {
         self.inner.groups.clone()
