+++ crates/pyext/src/operators/fermion_operator.rs
@@ -785,7 +785,7 @@ impl PyFermionOperator {
 
     /// An optional vector of `group indices` for each term.
     ///
-    /// For more information refer to the :mod:`~qiskit_fermions.operators.grouping` module.
+    /// For more information refer to the :mod:`~qiskit_fermions.operators.terms.grouping` module.
     #[getter]
     pub fn get_groups(&self) -> Option<Vec<u32>> {
         self.inner.groups.clone()
