+++ docs/cdoc/index.h
@@ -23,7 +23,11 @@
  */
 
 /**
- * @defgroup qf_operator_grouping qf_operator_grouping
+ * @defgroup qf_operator_terms qf_operator_terms
+ */
+
+/**
+ * @defgroup qf_operator_terms_grouping qf_operator_terms_grouping
  */
 
 /**
