+++ docs/cdoc/index.rst
@@ -12,8 +12,8 @@ Operators
    :maxdepth: 1
 
    qf-operators
-   qf-operators-grouping
    qf-operators-library
+   qf-operators-terms
 
 
 *******
