+++ docs/cdoc/qf-operators-terms-grouping.rst
@@ -1,0 +1,8 @@
+==================
+Grouping Functions
+==================
+
+.. doxygengroup:: qf_operator_terms_grouping
+   :content-only:
+   :members:
+   :undoc-members:
