+++ docs/cdoc/qf-operators-terms.rst
@@ -1,14 +1,26 @@
-.. _qf_operator_grouping:
+.. _qf_operator_terms:
 
-=================
-Operator Grouping
-=================
+==============
+Operator Terms
+==============
 
-Please refer to :ref:`grouping_explanation` for a detailed explanation of this
+.. toctree::
+   :hidden:
+   :maxdepth: 1
+
+   qf-operators-terms-grouping
+
+The C API provides various functions to operator on the individual terms of an operator,
+partitioning them based on their structure.
+
+Grouping
+--------
+
+Please refer to :ref:`grouping_explanation` for a detailed explanation of this module's
 functionality.
 
 Library
--------
+^^^^^^^
 
 Rather than always relying on the user to provide the group indices themselves,
 the C API provides a collection of functions which determine the grouping
@@ -20,7 +32,5 @@ information automatically.
   :c:func:`qf_group_terms_by_electronic_structure` Groups the terms of an operator by their electronic structure.
   ================================================ ==============================================================
 
-----
-
-.. doxygengroup:: qf_operator_grouping
+.. doxygengroup:: qf_operator_terms
    :content-only:
