+++ docs/pydoc/index.rst
@@ -12,8 +12,8 @@ Operators
    :maxdepth: 1
 
    qiskit_fermions.operators
-   qiskit_fermions.operators.grouping
    qiskit_fermions.operators.library
+   qiskit_fermions.operators.terms
 
 
 *******
