+++ docs/pydoc/qiskit_fermions.operators.terms.rst
@@ -1,4 +1,4 @@
-.. automodule:: qiskit_fermions.operators.grouping
+.. automodule:: qiskit_fermions.operators.terms
    :no-members:
    :no-inherited-members:
    :no-special-members:
